Add the Tailwind full config file
Add the Tailwind config file with the whole default
configuration.
On Tailwind it's needed to specify certain fields
on the config file to include determinated utilities
on the stylesheet such as the colors or fonts, you can
generate this file by using the `npx tailwindcss init -f`
command.
Therefore, there are two versions that you can generate,
the empty one which you can customize as much as you want,
and the default one which contains the whole Tailwind Css
configuration, but you can also costumize it if you'd like.