Modify the scripts and add some modules
The scripts on the package.json file were modified
in order of having a better experience when developing.
The changes applied were:
1.- The PostCss and Tailwind scripts suffered a change of
name, now they're called build and dev respectively.
2.- The scripts mentioned before don't have the --watch
flag anymore, instead, you have the option of use the
build:watch script.
Also, the postcss-nested plugin was added in order of
having the nesting feature available in the project.