Update the README file
The following changes were applied to the
README file:
1.- The Tailwind v3.0.0 section was added in order
of aware the lector that Tailwind has been updated
while writing the file and some things wrote before
may have troubles with the new version.
2.- The Typography section was removed, this was because
the section didn't contribute to the learning process as
much as I'd like, the section will be rewritten in the
future in a better way.